+++ mycli/main.py
@@ -43,7 +43,7 @@ from .packages.special.favoritequeries import FavoriteQueries
 from .sqlcompleter import SQLCompleter
 from .clitoolbar import create_toolbar_tokens_func
 from .clistyle import style_factory, style_factory_output
-from .sqlexecute import FIELD_TYPES, SQLExecute
+from .sqlexecute import FIELD_TYPES, SQLExecute, ERROR_CODE_ACCESS_DENIED
 from .clibuffer import cli_is_multiline
 from .completion_refresher import CompletionRefresher
 from .config import (write_default_config, get_mylogin_cnf_path,
@@ -432,7 +432,7 @@ class MyCli(object):
                     ssh_password, ssh_key_filename, init_command
                 )
             except OperationalError as e:
-                if ('Access denied for user' in e.args[1]):
+                if e.args[0] == ERROR_CODE_ACCESS_DENIED:
                     new_passwd = click.prompt('Password', hide_input=True,
                                               show_default=False, type=str, err=True)
                     self.sqlexecute = SQLExecute(